Check and power up M7MIX, and enable M7 TCM retention
M7MIX may be powered up by SYSROM if the boot container
includes M7 image which needs to be loaded into M7 TCM.
In other cases, such as M7 image loaded into DDR,
M7MIX is not powered on. The changes keep the M7MIX power
status as it was before OEI being executed. Aside of this
M7 TCM retention is turned on.

Signed-off-by: Florin Leotescu <florin.leotescu@nxp.com>