ddr: optimize training data size loaded from boot media
For stream devices such as USB the stream must point
to the following image body when OEI returns control to
ROM, therefore the loaded data size must be the entire
space allocated for DDR training data.
For all other devices the optimal size of loaded data
is the size of ddrphy_qb_state structure.

Signed-off-by: Viorel Suman <viorel.suman@nxp.com>